feat: replace RenderCacheProvider with new ServerRequestProvider cache
