--- packages/hydrogen/src/foundation/RenderCacheProvider/RenderCacheContext.tsx
@@ -1,6 +1,0 @@
-import {createContext} from 'react';
-import type {RenderCacheProviderProps} from './types';
-
-export const RenderCacheContext = createContext<RenderCacheProviderProps>({
-  cache: {},
-});
