--- packages/hydrogen/src/foundation/RenderCacheProvider/index.ts
@@ -1,1 +1,0 @@
-export {RenderCacheProvider} from './RenderCacheProvider';
