+++ packages/hydrogen/src/foundation/ServerRequestProvider/types.ts
@@ -1,0 +1,13 @@
+export type RequestCacheResult<T> =
+  | RequestCacheResultSuccess<T>
+  | RequestCacheResultError;
+
+type RequestCacheResultSuccess<T> = {
+  data: T;
+  error?: never;
+};
+
+type RequestCacheResultError = {
+  data?: never;
+  error: Response;
+};
