+++ packages/hydrogen/src/foundation/useQuery/hooks.ts
@@ -7,9 +7,8 @@ import {
   setItemInCache,
 } from '../../framework/cache';
 import {runDelayedFunction} from '../../framework/runtime';
-import {useRenderCacheData} from '../RenderCacheProvider/hook';
+import {useRequestCacheData} from '../ServerRequestProvider';
 
-import type {RenderCacheResult} from '../RenderCacheProvider/types';
 export interface HydrogenUseQueryOptions {
   cache: CacheOptions;
 }
@@ -26,8 +25,8 @@ export function useQuery<T>(
   queryFn: () => Promise<T>,
   /** Options including `cache` to manage the cache behavior of the sub-request. */
   queryOptions?: HydrogenUseQueryOptions
-): RenderCacheResult<T> {
-  return useRenderCacheData<T>(
+) {
+  return useRequestCacheData<T>(
     key,
     cachedQueryFnBuilder(key, queryFn, queryOptions)
   );
